Split out the handlers into a separate file
