+++ lambda-js/package.json
@@ -10,7 +10,7 @@
   },
   "author": "",
   "license": "ISC",
-  "dependencies": {
+  "devDependencies": {
     "aws-sdk": "^2.739.0"
   }
 }
