+++ lambda-js/step-lambda.js
@@ -1,7 +1,9 @@
-const AWS = require('aws-sdk');  // this sdk supports promises
+const AWS = require('aws-sdk')  // this sdk supports promises
 const tablename = process.env.TABLE_NAME
 const region = process.env.AWS_REGION
 const documentClient = new AWS.DynamoDB.DocumentClient({'region': region})
+const handlePostRequest = require("./handlers").handlePostRequest
+const handleGetRequest = require("./handlers").handleGetRequest
 
 // Importing aws-sdk using JS, believe document client needed for JS lambda writing
 // Could hardcode the table name, but bettter practice to put in environment variables (this is automatically done for AWS_REGION, and ACCESS_KEY)
@@ -28,9 +30,9 @@ exports.handler = async function(event, context) {
   try {
     switch(method) {
       case "POST":
-        return await handlePostRequest(event) //we are awaiting as this parent function returns the promise returned my the db update method
+        return await handlePostRequest(documentClient, tablename, event) //we are awaiting as this parent function returns the promise returned my the db update method
       case "GET":
-        return await handleGetRequest() //likewise as above
+        return await handleGetRequest(documentClient, tablename) //likewise as above
       default:
         return {"statusCode": 405, "body": "Please update lambda to handle this method"}
     }
@@ -43,63 +45,3 @@ exports.handler = async function(event, context) {
   }
 
 }
-
-// const handlePostRequest = async function(event) { << example of Functional Programming
-async function handlePostRequest(event) {
-  const body = JSON.parse(event.body) // converts a JSON string to a JSON object
-  console.log("body:", JSON.stringify(body)) // when logging log the string version otherwise will just say Object
-
-  const request = {
-    TableName: tablename,
-    Key:{
-      'Id': body.id
-    },
-    UpdateExpression: "ADD SeenCount :step",
-    ExpressionAttributeValues:{
-      ':step': body.step || 1,
-    },
-    ReturnValues:"UPDATED_NEW"
-  }
-  console.log("Calling DDB", JSON.stringify(request))
-
-  const result = await documentClient.update(request).promise()
-  console.log("result:", result)
-  return {"statusCode": 201, "body": JSON.stringify(result)}
- 
-}
-
-async function handleGetRequest(event) {
-
-  // nesting functions ok for recursive functions
-  async function dynamoScan(req, key, accumulator = []) { // default empty array used for base case
-    console.log("Dynamo Scan", JSON.stringify(req), JSON.stringify(key), accumulator.length)
-    // make a call to dynamo - adding the key to the request if it exists
-    const data = await documentClient.scan({...req, ExclusiveStartKey: key}).promise()
-    const newAccumulator = [...accumulator, ...data.Items]
-
-    // are we done? if so return
-    if (typeof data.LastEvaluatedKey == "undefined") { // base case
-      return newAccumulator
-    }
-    // recurse if not done
-    // to be able to keep going we need to pass in the LEK, and we need to amass the items
-    else {
-      return dynamoScan(req,data.LastEvaluatedKey, newAccumulator)
-    }
-    
-  }
-
-  const request = {
-    TableName: tablename,
-    Limit: 5
-  }
-  console.log("Scanning DDB", JSON.stringify(request))
-
-  const data = await dynamoScan(request)
-  // Code below will execute if promise resolves successfully, if want to respond to errors use try catch
-  console.log("Scan succeeded.");
-  data.map(item => console.log(item.Id + ": ",item.SeenCount)) // map better than for each
-  return {"statusCode": 200, "body": JSON.stringify(data)} // body should be a string representation of the status code
-
-}
-
